kafka-mysql-consumer/requirements.txt
kafka-python==2.0.2
mysql-connector-python==8.1.0
pandas==2.0.3
